Change to more unified package handling
Software packages (.deb) are not installed individually anymore but by
providing the .pkg file as it is available from CCO.  This way, all the
CML relevant Debian packages  (CML, PaTTY, IOL tools) will be pulled
from the software distribution package.  The only downside right now is
that the package stored in cloud storage is slightly bigger than the sum
of the actually required Debian packages as the .pkg usually includes
additional packages for upgrades (like a new kernel and sucht).

In addition:

- remove patty and iol customization scripts
- fix customize script so that it works witch changed hostnames
- change config vars in common and app section of config.yml
  - provide common.enable_patty boolean flag
  - rename app.deb to app.software
- remove unused / commented-out code blocks